VI: Move BufferQueue and NVFlinger to their own folder/namespace.
