+++ src/core/hle/service/vi/vi_m.h
@@ -8,6 +8,10 @@
 #include "core/hle/service/service.h"
 
 namespace Service {
+namespace NVFlinger {
+class NVFlinger;
+}
+
 namespace VI {
 
 class VI_M final : public ServiceFramework<VI_M> {
@@ -18,7 +22,7 @@ public:
 private:
     void GetDisplayService(Kernel::HLERequestContext& ctx);
 
-    std::shared_ptr<NVFlinger> nv_flinger;
+    std::shared_ptr<NVFlinger::NVFlinger> nv_flinger;
 };
 
 } // namespace VI
